Fix gcc 3.1 problems and warnings

+++ src/Packages/Uintah/CCA/Components/MPM/PhysicalBC/MPMPhysicalBCFactory.cc
@@ -35,7 +35,7 @@ void MPMPhysicalBCFactory::create(const ProblemSpecP& ps)
 
 void MPMPhysicalBCFactory::clean()
 {
-  for (int i = 0; i < mpmPhysicalBCs.size(); i++)
+  for (int i = 0; i < static_cast<int>(mpmPhysicalBCs.size()); i++)
     delete mpmPhysicalBCs[i];
   mpmPhysicalBCs.clear();
 }
